backend/requirements.txt
# Django y REST Framework
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0

# Seguridad y Autenticación
PyJWT==2.9.0
pyotp==2.9.0
django-otp==1.3.0
django-two-factor-auth==1.15.5

# Base de datos y ORM
sqlparse==0.5.3
psycopg2-binary==2.9.9  # Para PostgreSQL (opcional)
mysqlclient==2.2.4  # Para MySQL

# Utilidades
python-dotenv==1.0.1
Pillow==10.2.0  # Para manejo de imágenes
django-filter==24.1
django-cleanup==8.0.0

# Desarrollo y Testing
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.3
black==24.2.0
flake8==7.0.0

# Dependencias del sistema
asgiref==3.8.1
tzdata==2025.2
whitenoise==6.6.0  # Para servir archivos estáticos
gunicorn==21.2.0  # Para producción

# API y Documentación
drf-yasg==1.21.7  # Swagger/OpenAPI
markdown==3.5.2
